REW-7: Add view

+++ src/main/java/com/nester/Rew/controller/view/ApartmentWebController.java
@@ -4,11 +4,14 @@ import com.nester.Rew.service.ApartmentService;
 import com.nester.Rew.service.dto.apartment.ApartmentDto;
 import com.nester.Rew.service.dto.apartment.ApartmentDtoForSave;
 import com.nester.Rew.service.dto.apartment.ApartmentDtoForUpdate;
+import com.nester.Rew.service.dto.user.UserDto;
+import com.nester.Rew.service.impl.UserAppDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.data.web.SortDefault;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +19,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.util.List;
 
 @Controller
 @RequestMapping("/apartments")
@@ -27,47 +32,76 @@ public class ApartmentWebController {
 
     @GetMapping("/create")
     public String createForm() {
-        return "create_apartment";
+        return "apartment/create_apartment";
     }
 
     @PostMapping("/create")
     public String create(@ModelAttribute("apartment") ApartmentDtoForSave dto) {
+        UserAppDetails userAppDetails = (UserAppDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+        String email = userAppDetails.getUsername();
+        UserDto owner = new UserDto();
+        owner.setEmail(email);
+        dto.setOwner(owner);
         ApartmentDto created = apartmentService.create(dto);
         return "redirect:/apartments/" + created.getId();
     }
 
     @GetMapping
-    public String getAll(Model model, @PageableDefault(size = SIZE_PAGE) @SortDefault(SORT_PAGE) Pageable pageable) {
-        Page<ApartmentDto> apartments = apartmentService.getAll(pageable);
+    public String getAllActive(Model model, @PageableDefault(size = SIZE_PAGE) @SortDefault(SORT_PAGE) Pageable pageable) {
+        Page<ApartmentDto> apartments = apartmentService.getAllActive(pageable);
         model.addAttribute("apartments", apartments);
-        return "apartments";
+        return "apartment/apartments";
+    }
+
+    @GetMapping("/personal")
+    public String getAllByUser(Model model) {
+        UserAppDetails userAppDetails = (UserAppDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+        String email = userAppDetails.getUsername();
+        List<ApartmentDto> apartments = apartmentService.getAllByUser(email);
+        model.addAttribute("apartments", apartments);
+        return "apartment/personal_apartments";
     }
 
     @GetMapping("/{id}")
     public String getById(@PathVariable Long id, Model model) {
         ApartmentDto apartment = apartmentService.getById(id);
         model.addAttribute("apartment", apartment);
-        return "apartment";
+        return "apartment/apartment";
     }
 
     @GetMapping("/update/{id}")
     public String updateForm(@PathVariable Long id, Model model) {
         ApartmentDto toUpdate = apartmentService.getById(id);
         model.addAttribute("apartment", toUpdate);
-        return "update_apartment";
+        return "apartment/update_apartment";
     }
 
     @PostMapping("/update/{id}")
     public String update(@PathVariable(value = "id") Long id,
                          @ModelAttribute("apartment") ApartmentDtoForUpdate dto) {
-        dto.setId(id);
-        apartmentService.update(dto);
-        return "redirect:/apartments";
+        UserAppDetails userAppDetails = (UserAppDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+        String email = userAppDetails.getUsername();
+        if (apartmentService.getById(id).getOwner().getEmail().equals(email)) {
+            dto.setId(id);
+            UserDto owner = new UserDto();
+            owner.setEmail(email);
+            dto.setOwner(owner);
+            apartmentService.update(dto);
+        }
+        return "redirect:/apartments/" + id;
     }
 
     @GetMapping("/delete/{id}")
     public String delete(@PathVariable(value = "id") Long id) {
-        apartmentService.delete(id);
+        UserAppDetails userAppDetails = (UserAppDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+        String email = userAppDetails.getUsername();
+        if (apartmentService.getById(id).getOwner().getEmail().equals(email)) {
+            apartmentService.delete(id);
+        }
         return "redirect:/apartments";
     }
 }
